requirements.txt
pyqt5==5.15.3
pyyaml==5.4.1
